흠.. 생각날때마다.

+++ src/main/java/com/khc/study/infrastructure/dao/PostDao.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * PostDao
  */
-public interface PostDao extends JpaRepository<Post, Long> {
+public interface PostDao extends JpaRepository<Post, Integer> {
 
 }
